Fix scalar field (#45)
* fix scalar field

* remove cross import
