+++ src/napari/layers/image/_image_constants.py
@@ -137,3 +137,21 @@ VOLUME_DEPICTION_TRANSLATION = OrderedDict(
         (VolumeDepiction.PLANE, trans._('plane')),
     ]
 )
+
+
+class ImageProjectionMode(StringEnum):
+    """
+    Projection mode for aggregating a thick nD slice onto displayed dimensions.
+
+        * NONE: ignore slice thickness, only using the dims point
+        * SUM: sum data across the thick slice
+        * MEAN: average data across the thick slice
+        * MAX: display the maximum value across the thick slice
+        * MIN: display the minimum value across the thick slice
+    """
+
+    NONE = auto()
+    SUM = auto()
+    MEAN = auto()
+    MAX = auto()
+    MIN = auto()
